+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for materials and layers through the `Material` and `LayerStack` classes.
 - Added `Floor`, `Opening`, and `Roof` classes.
 - Added `IfcElementType` and `IfcPredefinedType` enumerations.
-- Added `BatchUpdate` context manager.
+- Added `Transaction` context manager.
 
 ### Changed